Add Asset class, ISearch interface. Update Search class to support handling exceptions.
